edit into more OOP format
