+++ src/main/java/Storage.java
@@ -2,21 +2,30 @@ import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Scanner;
+
 
 /**
  * Handles file operations such as ensuring file existence,
  * reading from a file, writing to a file, and loading tasks.
  */
-public class FileStoringRetrieving {
+public class Storage {
+
+    private String filePath;
+
+    public Storage(String filePath) {
+        this.filePath = filePath;
+        ensureFileExists();
+    }
+
 
 
     /**
      * Ensures that the specified file exists. If not, it creates the necessary directories and file.
      *
-     * @param filePath The path of the file to check and create if necessary.
      */
-    public static void ensureFileExists(String filePath) {
+    public void ensureFileExists() {
 
         File f = new File(filePath);
         if (!f.exists()) {
@@ -44,7 +53,7 @@ public class FileStoringRetrieving {
      * @param filePath The path of the file to read.
      * @throws FileNotFoundException If the file does not exist.
      */
-    public static void printFileContents(String filePath) throws FileNotFoundException {
+    public void printFileContents(String filePath) throws FileNotFoundException {
         File f = new File(filePath); // create a File for the given file path
         Scanner s = new Scanner(f); // create a Scanner using the File as the source
         while (s.hasNext()) {
@@ -55,10 +64,11 @@ public class FileStoringRetrieving {
     /**
      * Loads tasks from a file and stores them in the provided ArrayList.
      *
-     * @param filePath The path of the file to read from.
-     * @param storage  The ArrayList to store the loaded tasks.
+     * @return storage an {@code ArrayList} of {@code Task} objects loaded from the file.
+     *
      */
-    public static void loadTasksFromFile(String filePath, ArrayList<Task> storage) {
+    public ArrayList<Task> loadTasksFromFile() {
+        ArrayList<Task> storage = new ArrayList<>();
         File file = new File(filePath);
         try {
             Scanner scanner = new Scanner(file);
@@ -97,9 +107,11 @@ public class FileStoringRetrieving {
                 }
 
             }
+
         } catch (FileNotFoundException e) {
             System.out.println("File not found.");
         }
+        return storage;
     }
 
 
@@ -107,11 +119,10 @@ public class FileStoringRetrieving {
     /**
      * Writes tasks from an ArrayList to a file.
      *
-     * @param filePath The path of the file to write to.
      * @param storage  The ArrayList of tasks to save.
      * @throws IOException If an error occurs while writing to the file.
      */
-    public static void writeToFile(String filePath, ArrayList<Task> storage) throws IOException {
+    public void writeToFile(ArrayList<Task> storage) throws IOException {
         FileWriter fw = new FileWriter(filePath);
         for (Task task : storage) {
             if (task instanceof Todo) {
